Moved list by twitter handle endpoint
